APIO-116 Move ThrowableFunction to `throwable` package
